@@ -10,13 +10,13 @@ fn main() -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>> {
     task::block_on(async {
         let config = EnvironmentConfig::from_env()?;
         let endpoint = api::Features::endpoint(&config.api_url);
-        let client: http::HTTP<surf::Client> = http::HTTP::new(
+        let client: http::HTTP<reqwest::Client> = http::HTTP::new(
             config.app_name,
             config.instance_id,
             Uuid::new_v4().to_string(),
             config.secret,
         )?;
-        let res: api::Features = client.get(&endpoint).recv_json().await?;
+        let res: api::Features = client.get(&endpoint).send().await?.json().await?;
         dbg!(res);
         Ok(())
     })
